Unit Tests of ShoppingService hve been completed

+++ src/main/java/org/kodluyoruz/mybank/model/entity/dto/ShoppingDto.java
@@ -5,7 +5,7 @@ import lombok.Data;
 import org.kodluyoruz.mybank.model.entity.CardType;
 @Data
 public class ShoppingDto {
-    private Long spending;
+    private Double spending;
     private CardType cardType;
     private Long cardId;
     private String cardPassword;
